refactor(routes): consolidate /dashboard → /inspections (one canonical list)
The sidebar "Inspections" link pointed at a thin, feature-poor status-grouped
list (routes/inspections.tsx) while the full workspace — stats, + New
Inspection, filters, tabs, search, batch — lived at /dashboard. Two pages,
same backend endpoint (api.inspections.dashboard), confusing IA.

- Move the rich routes/dashboard.tsx → routes/inspections.tsx; serve it at
  /inspections. Delete the thin list. Drop the /dashboard route entirely
  (pre-launch: no back-compat redirect).
- Repoint every /dashboard page link: post-login/setup/join redirects,
  home "/" index + server root redirect, SSO/TOTP/identity-switch redirects,
  Google-calendar OAuth callback, concierge review email, editor exit,
  inspection-hub / form-renderer / feature-disabled back-links, calendar
  new-inspection link, reports-redirect (→ /inspections?workflow=published),
  CommandPalette, BackButton, NewInspectionWizard action, PWA manifest.
- Left untouched: the /api/inspections/dashboard data endpoint, the separate
  /agent-dashboard route, /dashboard-columns, dashboard-* libs/components,
  and the Cloudflare-dashboard doc mention.
- Updated unit tests (route-metadata, identity-service, concierge-emails,
  dashboard-tabs/-workflow imports) + e2e specs + docs.

type-check 0/0, test:web 590/590, test:unit 2162 passed, DS/migrefs OK.
