+++ app/routes/reports-redirect.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "react-router";
 
 export function loader() {
-  return redirect("/dashboard?workflow=published", 301);
+  return redirect("/inspections?workflow=published", 301);
 }
 
 export default function ReportsRedirect() {
