+++ server/lib/email-templates/sample-data.ts
@@ -4,7 +4,7 @@ const EXAMPLES: Record<string, string> = {
   address: '123 Main St', propertyAddress: '123 Main St',
   reportUrl: 'https://app.example.com/r/abc123', signUrl: 'https://app.example.com/sign/abc123',
   verifyUrl: 'https://app.example.com/verify/abc123', confirmUrl: 'https://app.example.com/confirm/abc123',
-  reviewUrl: 'https://app.example.com/review/abc123', dashboardUrl: 'https://app.example.com/dashboard',
+  reviewUrl: 'https://app.example.com/review/abc123', dashboardUrl: 'https://app.example.com/inspections',
   viewUrl: 'https://app.example.com/messages/abc123', acceptUrl: 'https://app.example.com/accept/abc123',
   payUrl: 'https://app.example.com/invoice/abc123',
   inviteLink: 'https://app.example.com/join/abc123', resetLink: 'https://app.example.com/reset/abc123',
