+++ tests/e2e/inspection-edit-hotkeys.spec.ts
@@ -32,7 +32,7 @@ test.describe('Inspection Edit hotkeys (Sprint 1 A-1..A-9)', () => {
         await page.fill('input[name=email]',    EMAIL!);
         await page.fill('input[name=password]', PASSWORD!);
         await page.click('button[type=submit]');
-        await page.waitForURL('**/dashboard');
+        await page.waitForURL('**/inspections');
     });
 
     test('? opens keyboard HUD with all 5 rating rows', async ({ page }) => {
