+++ tests/e2e/subsystem-a-inspector-tools-dock.spec.ts
@@ -26,7 +26,7 @@ test.describe('InspectorTools FAB dock (subsystem A M15)', () => {
         await page.fill('input[name=email]',    EMAIL!);
         await page.fill('input[name=password]', PASSWORD!);
         await page.click('button[type=submit]');
-        await page.waitForURL('**/dashboard');
+        await page.waitForURL('**/inspections');
         await page.goto(`/inspections/${INSPECTION_ID}/edit`);
         await page.waitForSelector('[x-data*=inspectionEditor]');
     });
