+++ tests/e2e/subsystem-a-photo-studio.spec.ts
@@ -29,7 +29,7 @@ test.describe('PhotoStudio MVP (subsystem A M14)', () => {
         await page.fill('input[name=email]',    EMAIL!);
         await page.fill('input[name=password]', PASSWORD!);
         await page.click('button[type=submit]');
-        await page.waitForURL('**/dashboard');
+        await page.waitForURL('**/inspections');
         await page.goto(`/inspections/${INSPECTION_ID}/edit`);
         await page.waitForSelector('[x-data*=inspectionEditor]');
     });
