+++ tests/e2e/subsystem-a-speed-mode.spec.ts
@@ -27,7 +27,7 @@ test.describe('SpeedMode (subsystem A M10)', () => {
         await page.fill('input[name=email]',    EMAIL!);
         await page.fill('input[name=password]', PASSWORD!);
         await page.click('button[type=submit]');
-        await page.waitForURL('**/dashboard');
+        await page.waitForURL('**/inspections');
         await page.goto(`/inspections/${INSPECTION_ID}/edit`);
         await page.waitForSelector('[x-data*=inspectionEditor]');
     });
